Kaller på simulering før iverksett, ellers kommer vi få "oppdrag finnes allerede" fra oppdraget, som betyder att kjeden allerede finnes i oppdrag. Sånn att vi må simulere før vi iverksetter.
